Adjust resource loader interop and assorted fixes
- Update ResourceLoader interop to match updated game signatures
- Add a StaticVfx string wrapper
- Wrap LoadIconByID signature scan in a try/catch to tolerate missing sig
- Change territory change handlers to use uint and adjust lookups accordingly
- Fix ExcelWorldHelper region/RowId casts
- Use Svc.Objects.LocalPlayer for Dismounting, switch font asset to NotoSansCjkMedium
- update ModelSystem weapon load calls to new sig
- Simplify GetPlayers by using LINQ.OfType

+++ DalamudPlugin/RaidsRewritten/EncounterManager.cs
@@ -105,12 +105,12 @@ public sealed class EncounterManager(
         dalamud.Framework.Update -= OnFrameworkUpdate;
     }
 
-    private void OnTerritoryChanged(ushort obj)
+    private void OnTerritoryChanged(uint obj)
     {
         ActiveEncounter?.Unload();
         statusPartyListProcessor.Reset();
 
-        if (this.encounterMap.TryGetValue(obj, out var encounter))
+        if (this.encounterMap.TryGetValue((ushort)obj, out var encounter))
         {
             ActiveEncounter = encounter;
             encounter.RefreshMechanics();
