+++ DalamudPlugin/RaidsRewritten/EntityManager.cs
@@ -84,7 +84,7 @@ public sealed class EntityManager : IDalamudHook
         this.world.DeleteWith<Attack>();
     }
 
-    private void OnTerritoryChanged(ushort obj)
+    private void OnTerritoryChanged(uint obj)
     {
         ClearAllManagedEntities();
     }
